% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABPR.R
\docType{data}
\name{ABPR}
\alias{ABPR}
\title{Range map for Abies procera}
\format{A SpatialPolygonDataFrame.}
\source{
\url{www.usgs.gov}
}
\usage{
ABPR
}
\description{
Historical range map for Abies procera, from "Atlas of United States Trees" by Elbert L. Little, Jr.
}
\seealso{
Other example data: \code{\link{climdat.fut}},
  \code{\link{climdat.hist}}
}
\keyword{datasets}
