% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ABPRclim.R
\docType{data}
\name{ABPRclim}
\alias{ABPRclim}
\title{Climate data for Abies procera}
\format{A RasterBrick with 19 layers.
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{www.worldclim.org}
}
\usage{
ABPRclim
}
\description{
Historical climate dataset of 19 bioclimate variables (\url{www.worldclim.org}), cropped to the extent of occurrence for Abies procera.
}
\keyword{datasets}

