% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\docType{methods}
\name{biplot}
\alias{biplot}
\alias{biplot,cnfa,GLcenfa-method}
\title{biplot}
\usage{
biplot(x, global, ...)

\S4method{biplot}{cnfa,GLcenfa}(x, global, xax = 1, yax = 2, p = 0.99,
  n = 5, plot = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{cnfa} or \code{enfa} describing the
occupied habitat}

\item{global}{an object of class \code{GLcnfa} describing the global reference
habitat}

\item{...}{additional \code{plot} arguments}

\item{xax}{the column number of the x-axis}

\item{yax}{the column number of the y-axis}

\item{p}{the proportion of observations to include in the calculations of
the minimum convex polygons}

\item{n}{the number of projected variables to label}

\item{plot}{if \code{TRUE}, plot will be returned on function call}
}
\description{
Biplots of \code{cnfa} and \code{enfa} objects.
}
\examples{
mod1 <- cnfa(x = climdat.hist, s.dat = ABPR, field = "CODE")
glc <- GLcenfa(x = climdat.hist)
biplot(x = mod1, global = glc)

}
