% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENFA.R
\docType{package}
\name{CENFA-package}
\alias{CENFA-package}
\title{CENFA: Climate- and Ecological-Niche Factor Analysis}
\description{
The CENFA package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

