% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnfa-class.R
\docType{class}
\name{cnfa-class}
\alias{cnfa-class}
\title{cnfa-class}
\description{
An object of class \code{cnfa} is created from performing climate-niche factor analysis on species presence data using the \code{cnfa} function.
}
\section{Slots}{

\describe{
\item{\code{call}}{Original function call.}

\item{\code{nf}}{Number of kept specialization factors.}

\item{\code{mf}}{marginality factor. Vector that describes the location of the species Hutchinsonian niche relative to the global niche.}

\item{\code{marginality}}{Standardized magnitude of the marginality factor.}

\item{\code{sf}}{specialization factors. Vector of eigenvalues.}

\item{\code{specialization}}{The square root of the sum of \code{sf} divided by the length of \code{sf}.}

\item{\code{s.prop}}{Vector representing the proportion of specialization found in each factor.}

\item{\code{co}}{A data frame of variable loadings with p rows and nf + 1 columns.}

\item{\code{ras}}{Raster* object of transformed climate values, with number of layers equal to nf + 1.}

\item{\code{present}}{numeric. Number of raster cells in which species is present.}
}}

