% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnfamad-class.R
\docType{class}
\name{cnfamad-class}
\alias{cnfamad-class}
\title{cnfamad-class}
\description{
An object of class \code{cnfamad} is created from performing climate-niche factor analysis on species presence data using the \code{cnfa} function.
}
\section{Slots}{

\describe{
\item{\code{call}}{original function call}

\item{\code{nf}}{number of kept specialization factors}

\item{\code{m}}{marginalities}

\item{\code{s}}{specialization factors}

\item{\code{map}}{Raster map of habitat suitability}
}}

