% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnfamad.R
\docType{methods}
\name{cnfamad}
\alias{cnfamad}
\alias{print.cnfamad,}
\alias{show.cnfamad}
\alias{cnfamad,RasterBrick,SpatialPolygonsDataFrame-method}
\alias{cnfamad,GLcnfamad,SpatialPolygonsDataFrame-method}
\title{Climate-Niche Factor Analysis}
\usage{
cnfamad(climdat, speciesdat, ...)

\S4method{cnfamad}{RasterBrick,SpatialPolygonsDataFrame}(climdat, speciesdat,
  field, nf = 1, scale = FALSE)

\S4method{cnfamad}{GLcnfamad,SpatialPolygonsDataFrame}(climdat, speciesdat,
  field, nf = 1)
}
\arguments{
\item{climdat}{Raster* object, typically a brick or stack of climate raster layers}

\item{speciesdat}{SpatialPolygons* object detailing species presence or abundance}

\item{field}{field of \code{speciesdat} that specifies presence or abundance. This is equivalent to the \code{field} argument in the \code{raster} package.}

\item{nf}{integer. Specifies the number of specialization axes to keep after transformation.}

\item{scale}{logical. If \code{TRUE} then the values of the Raster* object will
be centered and scaled. Depending on the resolution of the climate data and the
extent of the study area, this can be quite time consuming. If running this
function for multiple species, it is recommended that the climate data be scaled beforehand.}

\item{sp.prj}{character. Spatial projection of species data.}
}
\value{
Returns an S4 object of class \code{cnfa} with the following components:

call original function call

tab a dataframe
}
\description{
Performs climate-niche factor analysis using climate raster data and species presence data in shapefile.
}
