% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departure-class.R
\docType{class}
\name{departure-class}
\alias{departure-class}
\title{departure-class}
\description{
An object of class \code{departure} is created from performing ecological-niche
factor analysis on species presence data using the \code{enfa} function.
}
\section{Slots}{

\describe{
\item{\code{call}}{Original function call}

\item{\code{df}}{departure factor}

\item{\code{departure}}{Magnitude of the departure factor}

\item{\code{ras}}{Raster* object of transformed climate values}

\item{\code{weights}}{Raster layer of weights used for departure calculation}
}}

