% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difRaster.R
\docType{methods}
\name{difRaster}
\alias{difRaster}
\alias{difRaster,Raster,Raster-method}
\title{difRaster}
\usage{
difRaster(x, y, scale = TRUE, parallel = FALSE, n)

\S4method{difRaster}{Raster,Raster}(x, y, scale = TRUE, parallel = FALSE, n)
}
\arguments{
\item{x}{Raster* object of historical climate values}

\item{y}{Raster* object of future climate values}

\item{scale}{logical. If \code{TRUE} then the values of \code{x} and
\code{y} will be centered and scaled by the means and sds of the historical
climate data}

\item{parallel}{logical. If \code{TRUE} then multiple cores are utilized}

\item{n}{numeric. Optional number of CPU cores to utilize for parallel processing}
}
\value{
Raster* object
}
\description{
Calculates the absolute differences between two Raster* objects.
}
\examples{
difRaster(x = climdat.hist, y = climdat.fut)

}
