% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difRasterBrick-class.R
\docType{class}
\name{difRasterBrick-class}
\alias{difRasterBrick-class}
\title{difRasterBrick-class}
\description{
An object of class \code{difRasterBrick} is created by the \code{\link{difRaster}}
function, and inherits from the \code{RasterBrick} class.
}
\section{Slots}{

\describe{
\item{\code{title}}{Character}

\item{\code{file}}{Object of class ".RasterFile"}

\item{\code{data}}{Object of class ".SingleLayerData" or ".MultipleLayerData"}

\item{\code{history}}{To record processing history, not yet in use}

\item{\code{legend}}{Object of class .RasterLegend, Default legend. Should store
preferences for plotting. Not yet implemented except that it stores the color
table of images, if available}

\item{\code{extent}}{Object of Extent-class}

\item{\code{ncols}}{Integer}

\item{\code{nrows}}{Integer}

\item{\code{crs}}{Object of class "CRS", i.e. the coordinate reference system. In
Spatial* objects this slot is called 'proj4string'}
}}

