% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enfa-class.R
\docType{class}
\name{enfa-class}
\alias{enfa-class}
\title{enfa-class}
\description{
An object of class \code{enfa} is created from performing ecological-niche
factor analysis on species presence data using the \code{enfa} function.
}
\section{Slots}{

\describe{
\item{\code{call}}{Original function call.}

\item{\code{mf}}{numeric. Named vector representing the marginality factor, describing
the location of the species niche relative to the global niche.}

\item{\code{marginality}}{numeric. Magnitude of the marginality factor \code{mf}.}

\item{\code{sf}}{numeric. Named vector representing the specialization factor,
equivalent to the eigenvalues of specialization.}

\item{\code{specialization}}{numeric. The square of the sum of eigenvalues, divided
by the length of \code{sf}.}

\item{\code{p.spec}}{numeric. Named vector representing the proportion of
specialization found on each factor.}

\item{\code{co}}{p x p matrix of standardized variable loadings.}

\item{\code{cov}}{p x p species covariance matrix.}

\item{\code{present}}{numeric. Number of raster cells in which species is present.}

\item{\code{ras}}{RasterBrick of transformed environmental values, with p layers.}
}}

