% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enfa.R
\docType{methods}
\name{enfa}
\alias{enfa}
\alias{enfa,Raster,SpatialPoints-method}
\alias{enfa,RasterBrick,SpatialPolygonsDataFrame-method}
\alias{print.enfa,}
\alias{show.enfa}
\title{Ecological-Niche Factor Analysis}
\usage{
enfa(ecodat, speciesdat, ...)

\S4method{enfa}{RasterBrick,SpatialPolygonsDataFrame}(ecodat, speciesdat, field,
  nf = 1, scale = FALSE, return_values = FALSE)

\S4method{enfa}{Raster,SpatialPoints}(ecodat, speciesdat, field,
  fun = "count", nf = 1, scale = FALSE)
}
\arguments{
\item{ecodat}{Raster* object, typically a brick or stack of ecological raster layers}

\item{speciesdat}{SpatialPolygons* object detailing species presence or abundance}

\item{field}{field of \code{speciesdat} that specifies presence or abundance. This is equivalent to the \code{field} argument in the \code{raster} package.}

\item{nf}{integer. Specifies the number of specialization axes to keep after transformation.}

\item{scale}{logical. If \code{TRUE} then the values of the Raster* object will
be centered and scaled. Depending on the resolution of the climate data and the
extent of the study area, this can be quite time consuming. If running this
function for multiple species, it is recommended that the climate data be scaled beforehand.}

\item{mf}{marginality factor. Vector that describes the location of the species Hutchinsonian niche.}

\item{marginality.}{Standardized magnitude of the marginality factor.}

\item{s}{specialization factors. Matrix of specializations.}

\item{speciality}{specialization. Mahalanobis distance of something.}

\item{ras}{Raster* object of transformed climate values, with number of layers equal to nf + 1.}
}
\value{
Returns an S4 object of class \code{enfa} with the following components:
}
\description{
Performs ecological-niche factor analysis using environmental raster data and species presence data in shapefile.
}

