% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GLcnfa-class.R
\docType{class}
\name{GLcnfa-class}
\alias{GLcnfa-class}
\title{global.cnfa-class}
\description{
An object of class \code{global.cnfa} is created from performing climate-niche factor analysis on species presence data using the \code{cnfa} function.
}
\section{Slots}{

\describe{
\item{\code{call}}{original function call}

\item{\code{nf}}{number of kept specialization factors}

\item{\code{m}}{marginalities}

\item{\code{s}}{specialization factors}

\item{\code{map}}{Raster map of habitat suitability}
}}

