% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLcnfa.R
\docType{methods}
\name{GLcnfa}
\alias{GLcnfa}
\alias{print.cnfa,}
\alias{show.cnfa}
\alias{GLcnfa,RasterBrick-method}
\title{Climate Niche Factor Analysis}
\usage{
GLcnfa(x, ...)

\S4method{GLcnfa}{RasterBrick}(x, scale = FALSE, filename = "",
  progress = TRUE, cores = 1)
}
\arguments{
\item{x}{Raster* object, typically a brick or stack of climate raster layers}

\item{scale}{logical. If \code{TRUE} then the values of the Raster* object will
be centered and scaled. Depending on the resolution of the climate data and the
extent of the study area, this can be quite time consuming.}

\item{progress}{logical. If \code{TRUE} then progress updates are printed.}
}
\value{
Returns an S4 object of class \code{cnfa} with the following components:

call original function call

tab a dataframe
}
\description{
Performs climate niche factor analysis using climate raster data and species presence data in shapefile.
}
