% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GLcnfamad-class.R
\docType{class}
\name{GLcnfamad-class}
\alias{GLcnfamad-class}
\title{global.cnfamad-class}
\description{
An object of class \code{global.cnfamad} is created from performing climate-niche factor analysis on species presence data using the \code{cnfamad} function.
}
\section{Slots}{

\describe{
\item{\code{call}}{original function call}

\item{\code{nf}}{number of kept specialization factors}

\item{\code{m}}{marginalities}

\item{\code{s}}{specialization factors}

\item{\code{map}}{Raster map of habitat suitability}
}}

