% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLdeparture-class.R
\docType{class}
\name{GLdeparture-class}
\alias{GLdeparture-class}
\title{GLdeparture-class}
\description{
An object of class \code{GLdeparture} is created by the \code{\link{GLdeparture}}
function. It is best used for making comparisons between
species in the same study area. It speeds up the computation of multiple
departures by calculating the global covariance matrix as a first step, which
can then be fed into the \code{\link{departure}} function as a first argument.
This saves the user from having to calculate the global covariance matrix for
each species, which can take quite a bit of time.
}
\section{Slots}{

\describe{
\item{\code{global_difras}}{Raster* object of absolute differences between historical
\code{x} and future \code{y} climate values}

\item{\code{cov}}{matrix. Global covariance matrix}
}}

