% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/global_sigma.R
\name{global_sigma}
\alias{global_sigma}
\title{Climate-Niche Factor Analysis}
\usage{
global_sigma(climdat)
}
\arguments{
\item{climdat}{Raster* object, typically a brick or stack of climate raster layers}
}
\description{
Performs climate-niche factor analysis using climate raster data and species presence data in shapefile.
}

