% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\docType{methods}
\name{map}
\alias{map}
\alias{map,RasterLayer-method}
\title{Contrast adjustments for RasterLayer plots}
\usage{
map(x, ...)

\S4method{map}{RasterLayer}(x, type = "linear", n, legend.args, ...)
}
\arguments{
\item{x}{a RasterLayer}

\item{...}{Additional arguments for raster::plot}

\item{type}{character. Possible values are "linear", "stretch", and "sd"}

\item{n}{number of standard deviations to include if \code{type = "sd"}}

\item{legend.args}{a list of arguments to pass on to adjust the plot legend.
See \code{\link[fields]{image.plot}} for possible options}
}
\description{
A plotting function that provides methods for improving the contrast
between values.
}
\details{
If \code{type = "stretch"}, a histogram equalization procedure will be
applied to the values of \code{x}. If \code{type = "sd"}, the values of
\code{x} will be scaled between values that fall between \code{n} standard
deviations of the mean.
}
\examples{
mdr <- climdat.hist[[1]]
map(mdr)
map(mdr, type = "stretch")
map(mdr, type = "sd", n = 2)

}
