% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slots.R
\name{mf}
\alias{mf}
\title{Climate Niche Factor Analysis}
\usage{
mf(x)
}
\arguments{
\item{x}{cnfa or enfa object}
}
\description{
Functions for extracting data from slots of objects of classes \code{cnfa} and \code{enfa}.
}
\examples{
mod1 <- cnfa(x = climdat.hist, s.dat = ABPR, field = "CODE")
mf(mod1)

}
