% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names}
\alias{names}
\alias{names,Raster-method}
\alias{names,RasterStack-method}
\alias{names<-,Raster-method}
\title{names}
\usage{
\S4method{names}{Raster}(x)

\S4method{names}{RasterStack}(x)

\S4method{names}{Raster}(x) <- value
}
\arguments{
\item{x}{Raster* object}

\item{value}{character (vector)}
}
\value{
Character
}
\description{
Get or set the names of the layers of a Raster* object
}
\details{
Names of raster layers.
}
\examples{
names(climdat.hist)

}
