% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parScale.R
\docType{methods}
\name{parScale}
\alias{parScale}
\alias{parScale,Raster-method}
\title{Efficient scaling of Raster* objects}
\usage{
parScale(x, center = FALSE, scale = FALSE, parallel = FALSE, n)

\S4method{parScale}{Raster}(x, center = FALSE, scale = FALSE,
  parallel = FALSE, n)
}
\arguments{
\item{x}{Raster* object}

\item{center}{logical or numeric. If \code{TRUE}, centering is done by
subtracting the layer means (omitting NAs), and if \code{FALSE}, no centering
is done. If \code{center} is a numeric vector with length equal to the
\code{nlayers(x)}, then each layer of \code{x} has the corresponding value
from center subtracted from it}

\item{scale}{logical or numeric. If \code{TRUE}, scaling is done by dividing
the (centered) layers of \code{x} by their standard deviations if center is
\code{TRUE}, and the root mean square otherwise. If scale is \code{FALSE},
no scaling is done. If scale is a numeric vector with length equal to
\code{nlayers(x)}, each layer of \code{x} is divided by the corresponding
value. Scaling is done after centering}

\item{parallel}{logical. If \code{TRUE} then multiple cores are utilized}

\item{n}{numeric. Optional number of CPU cores to utilize for parallel processing}
}
\value{
Returns a matrix with the same row and column names as the layers of
  \code{x}. If \code{y} is supplied, then the covariances between the layers
  of \code{x} and the layers of code{y} are computed.
}
\description{
\code{scale} efficiently calculates the covariance between different climate
and ecological variables, taking advantage of parallel processing and pulling
data into memory only as necessary. For large datasets with lots of variables,
calculating the covariance matrix rapidly becomes unwieldy, as the number of
calculations required grows quadratically with the number of variables.
}
\details{
This function is designed to work similarly to the
  \code{\link[stats]{cov}} and the \code{\link[raster]{layerStats}}
  functions, with two major differences. First, \code{covmat} allows you to
  calculate the covariance between two different Raster* objects, whereas
  \code{layerStats} does not. Second, \code{covmat} can (optionally) compute
  each element of the covariance matrix in parallel, offering a dramatic
  improvement in computation time for large Raster* objects.

  The raster layer of weights \code{w} should contain raw weights as values,
  and should \emph{not} be normalized so that \code{sum(w) = 1}. This is
  necessary for computing the sample covariance, whose formula contains
  \code{sum(w) - 1} in its denominator.
}
\examples{
mat <- parScale(x = climdat.hist)

}
\seealso{
\code{\link[stats]{cov}}, \code{\link[raster]{layerStats}}
}
