% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\docType{methods}
\name{predict,cnfa-method}
\alias{predict,cnfa-method}
\alias{predict,enfa-method}
\alias{predict,departure-method}
\alias{predict,vulnerability-method}
\title{Predict methods}
\usage{
\S4method{predict}{cnfa}(object, newdata, filename = "", ...)

\S4method{predict}{enfa}(object, newdata, filename = "", ...)

\S4method{predict}{departure}(object, filename = "", ...)

\S4method{predict}{vulnerability}(object, newdata, filename = "", ...)
}
\arguments{
\item{object}{model object}

\item{newdata}{optional new data}

\item{filename}{character optional filename}

\item{...}{additional arguments}
}
\description{
Make a RasterLayer with predictions from a fitted model object.
}
