% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_map.R
\name{sensitivity_map}
\alias{sensitivity_map}
\alias{print.cnfa,}
\alias{show.cnfa}
\title{Create a sensitivity map}
\usage{
sensitivity_map(cnfa, cores = 1, filename = "", ...)
}
\arguments{
\item{cnfa}{cnfa object.}

\item{cores}{numeric. Number of cores to use for calculation (optional).}

\item{filename}{character. Output filename (optional).}

\item{...}{Additional arguments for file writing as for \code{\link[raster]{writeRaster}}.}
}
\value{
RasterLayer.
}
\description{
Creates a sensitivity map of species habitat.
}
\details{
The values of the sensitivity raster are calculated by centering the habitat's climate data around \strong{m} and projecting onto the sensitivity factor \strong{s}, given by formula  |\bold{S} - \bold{m}|\bold{s}.
}
