% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-data.R
\docType{data}
\name{tree-data}
\alias{tree-data}
\alias{ABPR,}
\alias{QUGA,}
\alias{SESE}
\alias{ABPR
Historical range map for the noble fir (\emph{Abies procera}).}
\alias{ABPR}
\alias{QUGA
Historical range map for the garry oak (\emph{Quercus garryana}).}
\alias{QUGA}
\alias{SESE
Historical range map for the coast redwood (\emph{Sequoia sempervirens}).}
\alias{SESE}
\title{Range map for Abies procera}
\format{A SpatialPolygonDataFrame.}
\source{
\url{www.usgs.gov}
}
\usage{
ABPR

QUGA

SESE
}
\description{
Some example datasets of historical tree distributions, from "Atlas of United
States Trees" by Elbert L. Little, Jr.
}
\seealso{
\code{\link{climdat.hist}}, \code{\link{climdat.fut}}
}
\keyword{datasets}
