% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerability-class.R
\docType{class}
\name{vulnerability-class}
\alias{vulnerability-class}
\title{vulnerability-class}
\description{
An object of class \code{vulnerability} is created from a \code{cnfa} object
and a \code{dep} object.
}
\section{Slots}{

\describe{
\item{\code{call}}{Original function call.}

\item{\code{vf}}{vulnerability factor.}

\item{\code{vulnerability}}{Magnitude of the vulnerability factor.}

\item{\code{present}}{Number of cells in which species is present.}

\item{\code{ras}}{RasterLayer of vulnerability values.}
}}

