% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerability.R
\docType{methods}
\name{vulnerability}
\alias{vulnerability}
\alias{vulnerability,cnfa,departure-method}
\title{Climatic vulnerability}
\usage{
vulnerability(cnfa, dep, ...)

\S4method{vulnerability}{cnfa,departure}(cnfa, dep, parallel = FALSE, n,
  filename = "", ...)
}
\arguments{
\item{cnfa}{Object of class \code{cnfa}}

\item{dep}{Object of class \code{departure}}

\item{...}{Additional arguments for file writing as for \code{\link[raster]{writeRaster}}}

\item{parallel}{logical. If \code{TRUE} then multiple cores are utilized}

\item{n}{numeric. Number of cores to use for calculation (optional)}

\item{filename}{character. Output filename (optional)}
}
\value{
Returns an S4 object of class \code{vulnerability} with the following slots:
\describe{
  \item{call}{Original function call}
  \item{vf}{Vulnerability factor. Vector of length p that describes the amount of
   vulnerability in each climate variable}
  \item{vulnerability}{Magnitude of the vulnerability factor}
  \item{ras}{RasterLayer of climate vulnerability}
  \item{weights}{Raster layer of weights used for departure calculation}
}
}
\description{
Calculates the climatic vulnerability of a species using a \code{cnfa} and
\code{departure} object.
}
\details{
The values of the vulnerability raster are calculated by projecting onto
the departure factor \strong{d}, given by formula

  \eqn{\nu} = \bold{Fv}.
}
\examples{
mod1 <- cnfa(x = climdat.hist, s.dat = ABPR, field = "CODE")
dep <- departure(x.hist = climdat.hist, x.fut = climdat.fut, s.dat = ABPR)
vuln <- vulnerability(cnfa = mod1, dep = dep)

}
\seealso{
\code{\link{departure}}
}
