% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abieproc.R
\docType{data}
\name{ABPR}
\alias{ABPR}
\title{Range map for Abies procera}
\format{A SpatialPolygonDataFrame.
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{www.worldclim.org}
}
\usage{
ABPR
}
\description{
Historical range map for Abies procera.
}
\keyword{datasets}
