% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdat.fut.R
\docType{data}
\name{climdat.fut}
\alias{climdat.fut}
\title{Future climate data}
\format{A RasterBrick with 10 layers:
\describe{
  \item{MDR}{mean diurnal range (mean of monthly max temp - min temp)}
  \item{ISO}{isothermality (MDR/TAR * 100)}
  \item{TS}{temperature seasonality (sd monthly temp * 100)}
  \item{HMmax}{max temp of warmest month}
  \item{CMmin}{min temp of coldest month}
  \item{PWM}{precip of wettest month}
  \item{PDM}{precip of driest month}
  \item{PS}{precip seasonality (sd/mean monthly precip)}
  \item{PWQ}{precip of wettest quarter}
  \item{PDQ}{precip of driest quarter}
}}
\source{
\url{https://www.worldclim.org}
}
\usage{
climdat.fut
}
\description{
Future climate dataset of 10 bioclimate variables (\url{https://www.worldclim.org}).
Based on the MIROC5 GCM projections under the RCP8.5 scenario for 2050.
}
\seealso{
\code{\link{climdat.hist}}, \code{\link{tree-data}}
}
\keyword{datasets}
