% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnfa.R
\docType{methods}
\name{cnfa}
\alias{cnfa}
\alias{print.cnfa,}
\alias{show.cnfa}
\alias{cnfa,GLcenfa,Spatial-method}
\alias{cnfa,Raster,Spatial-method}
\alias{cnfa,Raster,sf-method}
\alias{cnfa,Raster,matrix-method}
\title{Climate-Niche Factor Analysis}
\usage{
cnfa(x, s.dat, ...)

\S4method{cnfa}{GLcenfa,Spatial}(x, s.dat, field, fun = "last",
  filename = "", ...)

\S4method{cnfa}{Raster,Spatial}(x, s.dat, field, fun = "last", scale = TRUE,
  filename = "", ...)

\S4method{cnfa}{Raster,sf}(x, s.dat, field, fun = "last", scale = TRUE,
  filename = "", ...)

\S4method{cnfa}{Raster,matrix}(x, s.dat, prj, field, fun = "last",
  scale = TRUE, filename = "", ...)
}
\arguments{
\item{x}{Raster* object, typically a brick or stack with p climate
raster layers, or a \code{GLcenfa} object}

\item{s.dat}{matrix, SpatialPolygons*, SpatialPoints*, or sf object indicating
species presence}

\item{...}{Additonal arguments for rasterizing as for \code{\link[raster]{rasterize}}.}

\item{field}{field of \code{s.dat} that specifies presence or abundance. This
is equivalent to the \code{field} argument in \code{\link[raster]{rasterize}}.}

\item{fun}{function or character. Determines what values to assign to cells
with multiple spatial features, similar to the \code{fun} argument in
\code{\link[raster]{rasterize}}.}

\item{filename}{character. Optional filename to save the Raster* output to
file. If this is not provided, a temporary file will be created for large \code{x}.}

\item{scale}{logical. If \code{TRUE} then the values of \code{x} will get
centered and scaled. Depending on the resolution of the climate data and
the extent of the study area, this can be quite time consuming. If running
this function for multiple species, it is recommended that the data be
scaled beforehand using the \code{\link{GLcenfa}} function.}

\item{prj}{character. proj4string denoting projection of species observation data.}
}
\value{
Returns an S4 object of class \code{cnfa} with the following components:
\describe{
  \item{call}{Original function call.}
  \item{mf}{Marginality factor. Vector of length p that describes the location
  of the species Hutchinsonian niche relative to the global niche.}
  \item{marginality}{Magnitude of the marginality factor.}
  \item{sf}{Sensitivity factor. Vector of length p that describes the amount of
   sensitivity for each climate variable.}
  \item{sensitivity}{Magnitude of the sensitivity factor.}
  \item{s.prop}{Vector of length p representing the amount of specialization
  found in each CNFA factor.}
  \item{co}{A p x p matrix describing the amount of marginality and specialization
   on each CNFA factor. (The marginality column is normalized.)}
  \item{present}{Number of raster cells in which species is present.}
  \item{ras}{RasterBrick of transformed climate values, with p layers.}
}
}
\description{
Performs climate-niche factor analysis using climate raster data and species
presence data.
}
\examples{
mod1 <- cnfa(x = climdat.hist, s.dat = ABPR, field = "CODE")

# using GLcenfa as an initial step
# for multi-species comparison

glc <- GLcenfa(x = climdat.hist)
mod2 <- cnfa(x = glc, s.dat = ABPR, field = "CODE")

# same results either way
all.equal(m.factor(mod1), m.factor(mod2))
all.equal(s.factor(mod1), s.factor(mod2))

}
\seealso{
\code{\link{GLcenfa}}, \code{\link{enfa}}
}
