% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departure-class.R
\docType{class}
\name{departure-class}
\alias{departure-class}
\title{departure-class}
\description{
An object of class \code{departure} is created by the \code{\link{departure}}
departure function, which quantifies the amount of change between historical
and future climate conditions inside a species' habitat.
}
\section{Slots}{

\describe{
\item{\code{call}}{Original function call}

\item{\code{df}}{Departure factor}

\item{\code{departure}}{Magnitude of the departure factor}

\item{\code{g.cov}}{historical global covariance matrix}

\item{\code{ras}}{Raster* object of transformed climate values}

\item{\code{weights}}{Raster layer of weights used for departure calculation}
}}

