% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departure-class.R
\docType{class}
\name{departure-class}
\alias{departure-class}
\title{departure-class}
\description{
An object of class \code{departure} is created from performing ecological-niche factor analysis on species presence data using the \code{enfa} function.
}
\section{Slots}{

\describe{
\item{\code{call}}{Original function call.}

\item{\code{df}}{departure factor.}

\item{\code{departure}}{Standardized magnitude of the marginality factor.}

\item{\code{present}}{Number of cells in which species is present.}

\item{\code{ras}}{Raster* object of transformed climate values, with number of layers equal to nf + 1.}
}}

