% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departure-class.R
\docType{class}
\name{departure-class}
\alias{departure-class}
\title{departure-class}
\description{
An object of class \code{departure} is created from performing ecological-niche factor analysis on species presence data using the \code{enfa} function.
}
\section{Slots}{

\describe{
\item{\code{call}}{marginality factor. Vector that describes the location of the species Hutchinsonian niche.}

\item{\code{departure}}{Standardized magnitude of the marginality factor.}

\item{\code{distances}}{specialization factors. Matrix of specializations.}

\item{\code{departure_ras}}{Raster* object of transformed climate values, with number of layers equal to nf + 1.}

\item{\code{present}}{Number of cells in which species is present.}
}}

