% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departure.R
\docType{methods}
\name{departure}
\alias{departure}
\alias{departure,difRasterBrick,missing,cnfa-method}
\alias{departure,difRasterBrick,missing,Spatial-method}
\alias{departure,Raster,Raster,cnfa-method}
\alias{departure,Raster,Raster,Spatial-method}
\title{Climatic departure}
\usage{
departure(x.hist, x.fut, s.dat, ...)

\S4method{departure}{difRasterBrick,missing,cnfa}(x.hist, s.dat, field,
  fun = "last", ...)

\S4method{departure}{difRasterBrick,missing,Spatial}(x.hist, s.dat, field,
  fun = "last", ...)

\S4method{departure}{Raster,Raster,cnfa}(x.hist, x.fut, s.dat, field,
  fun = "last", scale = TRUE, parallel = FALSE, n, ...)

\S4method{departure}{Raster,Raster,Spatial}(x.hist, x.fut, s.dat, field,
  fun = "last", scale = TRUE, parallel = FALSE, n, ...)
}
\arguments{
\item{x.hist}{Raster* object, typically a brick or stack of historical climate
raster layers, or an object of class
\code{\link[=difRasterBrick-class]{difRasterBrick}} (see Details)}

\item{x.fut}{Raster* object, future climate values with the same layers as x.hist}

\item{s.dat}{SpatialPolygons*, sf, or cnfa object detailing species presence}

\item{...}{Additional arguments for \code{\link[raster]{clusterR}}}

\item{field}{field of \code{s.dat} that specifies presence. This is
equivalent to the \code{field} argument of \code{raster::rasterize}. Options
are 'first', 'last' (default), and 'count'}

\item{fun}{function or character. Determines what values to assign to cells
with multiple spatial features, similar to the \code{fun} argument in
\code{\link[raster]{rasterize}}}

\item{scale}{logical. If \code{TRUE} then the values of \code{x.hist} and
\code{x.fut} will be centered and scaled by the means and sds of the historical
climate data}

\item{parallel}{logical. If \code{TRUE} then multiple cores are utilized}

\item{n}{numeric. Optional number of CPU cores to utilize for parallel processing}
}
\value{
Returns an S4 object of class \code{departure} with the following slots:
\describe{
  \item{call}{Original function call}
  \item{df}{Departure factor. Vector of length p that describes the amount of
   departure between future and historical conditions for each climate variable}
  \item{departure}{Magnitude of the departure factor}
  \item{ras}{RasterBrick of climate departures, with p layers}
  \item{weights}{Raster layer of weights used for departure calculation}
}
}
\description{
This function quantifies the amount of change between historical and future
climate conditions inside a species' habitat.
}
\details{
For comparisons of multiple species in the same study area, it is much more
 efficient to first construct a Raster* object of absolute differences between
 the historical and future values, so that the differences do not need to be
 recalculated for each species. This can be acheived with by passing \code{x.hist}
 and \code{x.fut} to the \code{difRaster} function, and then passing the
 results to the \code{departure} function.
}
\examples{
dep1 <- departure(x.hist = climdat.hist, x.fut = climdat.fut, s.dat = ABPR, field = "CODE")

# using difRaster as an initial step
# for multi-species comparison

dif.ras <- difRaster(x.hist = climdat.hist, x.fut = climdat.fut)
dep2 <- departure(x.hist = dif.ras, s.dat = ABPR, field = "CODE")

# same results either way
all.equal(dep1@df, dep2@df)

}
