% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/departure.R
\docType{methods}
\name{departure}
\alias{departure}
\alias{departure,RasterBrick,RasterBrick,SpatialPolygonsDataFrame-method}
\alias{departure,Raster,Raster,enfa-method}
\alias{departure,RasterBrick,RasterBrick,cnfa-method}
\alias{departure,RasterBrick,missing,cnfa-method}
\title{Climatic departure}
\usage{
departure(x.hist, x.fut, s.dat, ...)

\S4method{departure}{RasterBrick,RasterBrick,SpatialPolygonsDataFrame}(x.hist,
  x.fut, s.dat, field, scale = FALSE, depart.ras = TRUE)

\S4method{departure}{Raster,Raster,enfa}(x.hist, x.fut, s.dat,
  depart.ras = TRUE, scale = FALSE)

\S4method{departure}{RasterBrick,RasterBrick,cnfa}(x.hist, x.fut, s.dat,
  depart.ras = TRUE, scale = FALSE)

\S4method{departure}{RasterBrick,missing,cnfa}(x.hist, x.fut, s.dat,
  depart.ras = TRUE, scale = FALSE)
}
\arguments{
\item{x.hist}{Raster* object, typically a brick or stack of historical climate raster layers}

\item{x.fut}{Raster* object, future climate values with the same layers as x.hist}

\item{s.dat}{SpatialPolygons* object detailing species presence or abundance}

\item{field}{field of \code{speciesdat} that specifies presence or abundance. This is equivalent to the \code{field} argument of \code{raster::rasterize}.}

\item{scale}{logical. If \code{TRUE} then the values of \code{x.hist} and \code{x.fut} will
be centered and scaled by the means and sds of the historical climate data. Depending on the resolution of the data and the
extent of the study area, this can be quite time consuming. If running this
function for multiple species, it is recommended that the climate data be scaled beforehand.}

\item{nf}{integer. Specifies the number of specialization axes to keep after transformation.}

\item{sp.prj}{character. Spatial projection of species data.}
}
\value{
Returns an S4 object of class \code{cnfa} with the following slots:

call original function call

departure climatic departure D of species

departure_ras raster of distances d_i

present number of cells in which species is present
}
\description{
Calculates the climatic departure of a species using historical and future climate raster data and species presence data.
}
