% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difRaster.R
\docType{methods}
\name{difRaster}
\alias{difRaster}
\alias{difRaster,Raster,Raster-method}
\title{difRaster}
\usage{
difRaster(x.hist, x.fut, scale = TRUE, parallel = FALSE, n)

\S4method{difRaster}{Raster,Raster}(x.hist, x.fut, scale = TRUE,
  parallel = FALSE, n)
}
\arguments{
\item{x.hist}{Raster* object of historical climate values}

\item{x.fut}{Raster* object of future climate values}

\item{scale}{logical. If \code{TRUE} then the values of \code{x.hist} and
\code{x.fut} will be centered and scaled by the means and sds of the historical
climate data}

\item{parallel}{logical. If \code{TRUE} then multiple cores are utilized}

\item{n}{numeric. Optional number of CPU cores to utilize for parallel processing}
}
\value{
A Raster* object of class \code{difRasterBrick}
}
\description{
An object of class \code{enfa} is created from performing ecological-niche
factor analysis on species presence data using the \code{enfa} function.
}
