% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enfa-class.R
\docType{class}
\name{enfa-class}
\alias{enfa-class}
\title{enfa-class}
\description{
An object of class \code{enfa} is created from performing ecological-niche factor analysis on species presence data using the \code{enfa} function.
}
\section{Slots}{

\describe{
\item{\code{mf}}{marginality factor. Vector that describes the location of the species Hutchinsonian niche.}

\item{\code{marginality.}}{Standardized magnitude of the marginality factor.}

\item{\code{s}}{specialization factors. Matrix of specializations.}

\item{\code{specialization}}{specialization. Mahalanobis distance of something.}

\item{\code{ras}}{Raster* object of transformed climate values, with number of layers equal to nf + 1.}

\item{\code{present}}{Number of cells in which species is present.}
}}

