% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enfa.R
\docType{methods}
\name{enfa}
\alias{enfa}
\alias{print.enfa,}
\alias{show.enfa}
\alias{enfa,Raster,SpatialPolygonsDataFrame-method}
\alias{enfa,Raster,SpatialPoints-method}
\title{Ecological-Niche Factor Analysis}
\usage{
enfa(ecodat, speciesdat, ...)

\S4method{enfa}{Raster,SpatialPolygonsDataFrame}(ecodat, speciesdat, field,
  nf = 1, scale = FALSE, mar.type = "Basille", return_values = FALSE,
  ...)

\S4method{enfa}{Raster,SpatialPoints}(ecodat, speciesdat, field, nf = 1,
  scale = FALSE, fun = "count", mar.type = "Basille",
  return_values = FALSE, ...)
}
\arguments{
\item{ecodat}{Raster* object, typically a brick or stack of ecological raster layers}

\item{speciesdat}{SpatialPolygons* object detailing species presence or abundance}

\item{field}{field of \code{speciesdat} that specifies presence or abundance. This is equivalent to the \code{field} argument in the \code{raster} package.}

\item{nf}{integer. Specifies the number of specialization axes to keep after transformation.}

\item{scale}{logical. If \code{TRUE} then the values of the Raster* object will
be centered and scaled. Depending on the resolution of the climate data and the
extent of the study area, this can be quite time consuming. If running this
function for multiple species, it is recommended that the climate data be scaled beforehand.}

\item{mar.type}{character. Choices are "Hirzel" or "Basille". See details.}

\item{mf}{marginality factor. Vector that describes the location of the species Hutchinsonian niche.}

\item{marginality.}{Standardized magnitude of the marginality factor.}

\item{s}{specialization factors.  of specializations.}

\item{speciality}{specialization. Vector}

\item{ras}{Raster* object of transformed climate values, with number of layers equal to nf + 1.}
}
\value{
Returns an S4 object of class \code{enfa} with the following components:
}
\description{
Performs ecological-niche factor analysis using environmental raster data and species presence data in shapefile.
}
\details{
Hirzel et al. (2002) defined the overall marginality $M$ as a standardized distance between the centroid of the species' niche and the global niche, given by $$M = \frac{\sqrt{\sum\limits_{j=1}^P m_j^2}}{1.96}.$$ Basille and Calenge (2008), however, defined $M$ as $$M = \mathbf{m}^T\mathbf{m}$$. The default \code{mar.type} reflects Basille and Calenge's definition so that results will by default agree with those calculated using Basille and Calenge's \code{adehabitatHS} package.
}
