% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLcenfa.R
\docType{methods}
\name{GLcenfa}
\alias{GLcenfa}
\alias{print.GLcenfa,}
\alias{show.GLcenfa}
\alias{GLcenfa,RasterBrick-method}
\alias{GLcenfa,RasterStack-method}
\title{GLcenfa}
\usage{
GLcenfa(x, ...)

\S4method{GLcenfa}{RasterBrick}(x, scale = FALSE, filename = "",
  progress = TRUE, cores = 1, ...)

\S4method{GLcenfa}{RasterStack}(x, scale = FALSE, filename = "",
  progress = TRUE, cores = 1, ...)
}
\arguments{
\item{x}{Raster* object, typically a brick or stack of p environmental raster layers.}

\item{scale}{logical. If \code{TRUE} then the values of \code{x} will
be centered and scaled. Depending on the resolution of the data and the
extent of the study area, this can be quite time consuming.}

\item{filename}{character. Optional filename to save the RasterBrick output to file. If this is not provided, a temporary file will be created for large \code{x}.}

\item{progress}{logical. If \code{TRUE} then progress updates are printed.}

\item{cores}{numeric. Number of cores to utilize for speedier parallel computation of the covariance matrix.}
}
\value{
Returns an S4 object of class \code{GLcenfa} with the following components:

global_ras Raster* brick \code{x} with p layers, possibly centered and scaled.

cov matrix. Global p x p covariance matrix.

center numeric. Layer means of \code{x}.

sd numeric. Layer standard deviations of \code{x}.

ncells numeric. Total number of raster cells with environmental data.
}
\description{
This function is used to facilitate comparisons between species in the same study area. It speeds up the computation of multiple CNFAs or ENFAs by calculating the global covariance matrix as a first step, which can then be fed into the \code{\link{cnfa}} or \code{\link{enfa}} functions as their first argument. This saves the user from having to calculate the global covariance matrix for each species, which can take quite a bit of time.
}
\seealso{
\code{\link{cnfa}}, \code{\link{enfa}}
}
