% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict,cnfa-method}
\alias{predict,enfa-method}
\alias{predict,departure-method}
\alias{predict,vulnerability-method}
\title{Predict methods}
\usage{
\S4method{predict}{cnfa}(object, newdata, filename = "", parallel = FALSE, n = 1, ...)

\S4method{predict}{enfa}(object, newdata, filename = "", parallel = FALSE, n = 1, ...)

\S4method{predict}{departure}(object, filename = "", parallel = FALSE, n = 1, ...)

\S4method{predict}{vulnerability}(object, newdata, filename = "", parallel = FALSE, n = 1, ...)
}
\arguments{
\item{object}{model object}

\item{newdata}{optional new data}

\item{filename}{character. Optional filename to save the RasterBrick output
to file. If this is not provided, a temporary file will be created for large
\code{x}}

\item{parallel}{logical. If \code{TRUE} then multiple cores are utilized}

\item{n}{numeric. Number of CPU cores to utilize for parallel processing}

\item{...}{Additional arguments for \code{\link[raster]{writeRaster}}}
}
\description{
Make a RasterLayer with predictions from a fitted model object.
}
