% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\docType{methods}
\name{raster,enfa-method}
\alias{raster,enfa-method}
\alias{raster,cnfa-method}
\alias{raster,GLcnfa-method}
\title{CNFA raster extraction}
\usage{
\S4method{raster}{enfa}(x)

\S4method{raster}{cnfa}(x)

\S4method{raster}{GLcnfa}(x)
}
\arguments{
\item{cnfa}{object of class \code{cnfa}}
}
\value{
Raster* object.
}
\description{
Extracts Raster* object from an object of class \code{cnfa}.
}
