% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\docType{methods}
\name{raster,enfa-method}
\alias{raster,enfa-method}
\alias{raster,cnfa-method}
\alias{raster,GLcenfa-method}
\title{CENFA raster extraction}
\usage{
\S4method{raster}{enfa}(x)

\S4method{raster}{cnfa}(x)

\S4method{raster}{GLcenfa}(x)
}
\arguments{
\item{x}{object of class \code{cnfa}, \code{enfa}, or \code{GLcenfa}}
}
\value{
Raster* object.
}
\description{
Extracts Raster* objects from objects of classes \code{\link{cnfa}}, \code{\link{enfa}}, and \code{\link{GLcenfa}}.
}
\examples{
mod1 <- cnfa(x = climdat.hist, s.dat = ABPR, field = "CODE")
ras <- raster(mod1)

}
