% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slots.R
\name{slot-access}
\alias{slot-access}
\alias{m.factor}
\alias{s.factor}
\alias{marginality}
\alias{specialization}
\alias{sensitivity}
\title{Accessing CENFA slots}
\usage{
m.factor(x)

s.factor(x)

marginality(x)

specialization(x)

sensitivity(x)
}
\arguments{
\item{x}{cnfa or enfa object}
}
\description{
Functions for extracting data from slots of objects of classes \code{cnfa} and \code{enfa}.
}
\examples{
mod1 <- cnfa(x = climdat.hist, s.dat = ABPR, field = "CODE")
m.factor(mod1)

}
