% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-data.R
\name{tree-data}
\alias{tree-data}
\alias{ABPR}
\alias{QUGA}
\alias{SESE}
\title{Tree distribution data}
\format{SpatialPolygonDataFrame}
\source{
\url{https://www.usgs.gov/}
}
\description{
Some example datasets of historical tree distributions, from "Atlas of United
States Trees" by Elbert L. Little, Jr.
}
\details{
\describe{
  \item{ABPR}{Historical range map for the noble fir (\emph{Abies procera})}
  \item{QUGA}{Historical range map for the Oregon white oak (\emph{Quercus garryana})}
  \item{SESE}{Historical range map for the coast redwood (\emph{Sequoia sempervirens})}
}
}
\seealso{
\code{\link{climdat.hist}}, \code{\link{climdat.fut}}
}
